CVS-71 Transaction Monthly Sum QueryDSL 적용

+++ src/main/java/com/alliex/cvs/web/dto/SettleTransMonthlySumRequest.java
@@ -11,15 +11,15 @@ public class SettleTransMonthlySumRequest {
 
     private Long userId;
 
-    private Long approvalCount;
+    private Integer approvalCount;
 
     private Long approvalAmount;
 
-    private Long cancelCount;
+    private Integer cancelCount;
 
     private Long cancelAmount;
 
-    public Long getTotalCount() {
+    public Integer getTotalCount() {
         return approvalCount + cancelCount;
     }
